+++ projects/fake_typing/fake_typing/fake.py
@@ -1,7 +1,19 @@
-class FakeType:
-
-    def __getattr__(self, _) -> 'FakeType':
-        return self    
-
-    def __getitem__(self, _) -> 'FakeType':
+class _FakeType:
+    
+    def __getattr__(self, _) -> '_FakeType':
         return self
+    
+    def __getitem__(self, _) -> '_FakeType':
+        return self
+    
+    def __call__(self) -> '_FakeType':
+        import warnings
+        warnings.warn(
+            '`FakeType()` is deprecated, use `FakeType` directly.',
+            DeprecationWarning
+        )
+        return self
+
+
+# singleton
+FakeType = _FakeType()