+++ projects/fake_typing/fake_typing/register.py
@@ -6,5 +6,5 @@ from .fake import FakeType
 
 def register_fake_type(name='T') -> Any:
     last_frame = currentframe().f_back
-    out = last_frame.f_globals[name] = FakeType()
+    out = last_frame.f_globals[name] = FakeType
     return out