+++ projects/fake_typing/changelog.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+### 0.1.3 (2024-07-02)
+
+- make `FakeType` singleton.
+- correct `FALSE_CASE`.
+
 ### 0.1.2 (2024-06-29)
 
 - add `FALSE_CASE` constant.